+++ changelog.md
@@ -3,6 +3,9 @@
 _Note: This changelog may be incomplete._
 
 ## Next version (2018-??-??)
+- meteor-chess-client: add output field to indicate when engine offers or claims
+  a draw (JSON output mode only), and show intention to claim (text output mode
+  only), respectively
 - simple-chess-engine: implement support for commands setboard and draw
 
 ## Version 0.4.1 (2018-02-09)